+++ src/containers/Register/index.js
@@ -3,7 +3,7 @@ import { connect } from 'react-redux';
 import { RegisterChoice, GoBack, FaceBook } from './register.components.js';
 import Internal from './Internal';
 import Login from '../Login';
-import { registerNewUser,loginUser } from '../../actions/users';
+import { registerNewUser } from '../../actions/users';
 
 
 
@@ -22,7 +22,6 @@ class Register extends Component {
 
     this.registrationChoice=this.registrationChoice.bind(this);
     this.registerUser=this.registerUser.bind(this);
-    this.loginUser=this.loginUser.bind(this);
   }
 
 
@@ -49,23 +48,7 @@ class Register extends Component {
        
   }
 
-  loginUser(email,password){
-    let local = { email: email, password: password }
-    this.props.loginUser(local); 
-    setTimeout(()=>{ //using this to wait for server response
-      const {users} = this.props;
-      if(users === 'success'){
-      this.setState({registrationChoice: 'loggedin'})
-      }
-      else if(users === 302){
-        this.setState({err: true, registrationChoice: 'login'})
-      }
-      else if(users === 400){
-        alert('server error');
-      }
-    },500)
-
-  }
+  
 
   
 
@@ -110,7 +93,7 @@ class Register extends Component {
         return (
           <div className="register-container">
           {err ? <div>User Not Found</div> : null }
-            <Login loginUser={this.loginUser} />
+            <Login />
           </div>
           )
         case "loggedin":
@@ -132,7 +115,7 @@ const mapStateToProps = (state) => {
 
 const ConnectedRegister = connect(
   mapStateToProps,
-  {registerNewUser,loginUser}
+  {registerNewUser}
 )(Register)
